First example for control file validation
The validation is done through a regular expression. Right now it only
checks that the path is a subpath of '/sys/'. This should not break any
current or future devices.

If the path is determined invalid, this is shown clearly - both in the
settings page and in the config file.

This is a first proposal for
https://github.com/sam-hobbs/harbour-hammerhead-torch/issues/3